+++ domain/src/main/java/com/weather/domain/service/WeatherClient.java
@@ -1,9 +1,7 @@
 package com.weather.domain.service;
 
-import java.math.BigDecimal;
-
 import com.weather.domain.model.WeatherResponse;
 
 public interface WeatherClient {
-	WeatherResponse retrieveWeather(BigDecimal latitude, BigDecimal longitude);
+	WeatherResponse retrieveWeather(Double latitude, Double longitude);
 }
